Fix MPCKMeans
- Key fix: don't reset labels between each iteration. Instead, keep cluster assignments
from previous round and just move points as-needed.
This is the behavior used by the code published by the authors at
https://www.cs.utexas.edu/users/ml/risc/code/ in MPCKMeans.java:runKMeans() where
m_ClusterAssignments is not reset between each iteration.

- Second modification: only points connected by a cannot-link constraint are considered
when finding the farthest pairs of points for f_c. This matches the behavior in the code
published by the authors in WeightedMahalanobis.java:getMaxPoints.

- The rest of the modifications include enabling specifying a random number generator
for repeatable runs, collecting more statistics about convergence, and enabling clustering
over more points than we have ground truth for by specifying the indices that the oracle
can ask about.